fix: Handle content-length mismatch + clean cancellation
This change fix a lot of reliability issues related that were discovered
when testing sending a mismatched content-length header from the client.
It caused the tunnel to hang as it was waiting for more data indefinitely
in the case where the content-length was larger than the actual data that was
sent.

Other issues were discovered when further requests were made while the tunnel
was in the blocking state, additionally when the mmar client was killed while
in that state.

This fix address all these issues, ensuring that we do not trust that header
coming from the user (client), adding the "connection: close" header, as well
has handling cleanly killing all the goroutines on the server associated with the
tunnel client when a mmar client is killed. Such as draining all the channels to
allow the goroutine to reach the return condition instead of blocking indefinitely.