fix: Remove client read message timeout
When a TunnelMessage that is large coming from the server
the Conn in ReceiveMessage would sometimes timeout on read.
This timeout was initially added to check if the client has been shutdown
gracefully, but that caused more issues than needed. So we fall back to
the client connection being shutdown from the server when the client is
shutdown by the user.