feat: Reconnecting mmar clients
This change adds the ability for the mmar client to attempt to reconnect
if there is ever a disconnection of the tunnel. Which allows for the
persistance of the subdomain as long as the mmar client is running.
Once the mmar client is shutdown and then restarted, a new subdomain
will be issued.