feat: Add heartbeats between server and clients
Adding heartbeats as a means to determine if the connection between
the client and the server is still alive and vice versa. This should
allow for better handling for states when they get disconnected,
additionally being able to determine if that is the case quicker.

This change updates the mmar Message Protocol.